feat(下载器): 添加分辨率解析功能并更新相关组件
- 新增解析器模块用于从文件名提取分辨率信息
- 将分辨率字段从u16改为u32以支持4K解析
- 在下载器相关组件中添加分辨率字段
- 更新前端组件以显示动态分辨率
- 缩短Bangumi API缓存时间

+++ src-tauri/src/models/mikan.rs
@@ -21,7 +21,7 @@ pub struct MikanResourceItem {
     #[ts(optional)]
     pub episode_range: Option<String>,
     #[ts(optional)]
-    pub resolution: Option<u16>,
+    pub resolution: Option<u32>,
     #[ts(optional)]
     pub subtitle_lang: Option<String>,
     #[ts(optional)]
